+++ postgres/CHANGES.md
@@ -13,6 +13,22 @@
     let row_affected: u64 = "SELECT 1; SELECT 1".execute(&client).await?;
     ```
 - remove `Client::pipeline` and `Pool::pipeline`. `pipeline::Pipeline` type can be execute with `Execute::query` method
+    
+  remove `pipeline::Pipeline::query` and `pipeline::Pipeline` can be queried with `ExecuteMut::query_mut` method
+    ```rust
+    use xitca_postgres::Execute;
+    // prepare statement and create a pipeline
+    let stmt = Statement::named("SELECT 1", &[]).execute(&client).await?;
+    let mut pipe = Pipeline::new();
+    
+    // use ExecuteMut trait to add query to pipeline
+    use xitca_postgres::ExecuteMut;
+    stmt.query(&mut pipe)?;
+    stmt.query(&mut pipe)?;
+
+    // use Execute trait to start pipeline query
+    let pipe_stream = pipe.query(&client)?;
+    ```
 - remove `dev::AsParams` trait export. It's not needed for implementing `Query` trait anymore    
 
 ## Change
@@ -32,21 +48,19 @@
     // statement have no value params and can be used for query.
     let stream = stmt.query(&client)?;
     ```
-- rename `Pipeline::query` to `Pipeline::pipe_query` to avoid naming collision with `Execute::query`   
 - `AsyncLendingIterator` is no longer exported from crate's root path. use `iter::AsyncLendingIterator` instead
 - `query::RowStreamOwned` and `row::RowOwned` are no longer behind `compat` crate feature anymore
 - `statement::Statement::unnamed` must bind to value parameters with `bind` or `bind_dyn` before calling `Execute` methods.
     ```rust
     let stmt = Statement::unnamed("SELECT * FROM users WHERE id = $1", &[Type::INT4]);
-    let row_stream = stmt.bind([9527]).query(&client);
+    let row_stream = stmt.bind([9527]).query(&client)?;
     ```
 - `Query::_send_encode_query` method's return type is changed to `Result<(<S as Encode>::Output, Response), Error>`. Enabling further simplify of the surface level API at the cost of more internal complexity
 - `Encode` trait implementation detail change
 - `IntoStream` trait is renamed to `IntoResponse` with implementation detail change
 
 ## Add
-- add `Execute` trait for extending query customization
-- add `Client::prepare_blocking`
+- add `Execute` and `ExecuteMut` traits for extending query customization
 - add `Prepare::{_prepare_blocking, _get_type_blocking}`
 - add `iter::AsyncLendingIteratorExt` for extending async iterator APIs
 - add `statement::Statement::{bind, bind_dyn}` methods for binding value parameters to a prepared statement for query